Finished fixing bugs

+++ UI.RichClient/LpcAsset.cs
@@ -51,7 +51,7 @@ namespace UI.RichClient
         public float CmdsPerSecond { get; set; }
         public string Sent { get; set; }
         public byte[] Color { get; set; }
-        public byte CurMode { get; set; }
+        public int CurMode { get; set; }
         public List<Mode> Modes { get { return _modes; } }
         public byte BoxNumber { get; set; }
         public int Delay { get; set; }
